requirements.txt
pytesseract
flask
pillow
numpy
opencv-contrib-python
uswgi
pdf2image